Blog List Expansion, step 9

+++ part4/bloglistbackend/controllers/blogs.js
@@ -40,8 +40,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const body = request.body
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const blog = await Blog.findById(request.params.id)
+
+  
+  if (decodedToken.id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(403).json({ error: 'You are not authorized to delete this blog' });
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
